requirements.txt
streamlit==1.32.0
streamlit-authenticator==0.2.3
pandas<2.0.0
numpy<2.0.0
matplotlib<3.8.0
plotly==5.18.0
ccxt==4.1.13
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
boto3==1.34.34
PyYAML==6.0.2
bcrypt==4.0.1
requests==2.32.3
